Detect interface changes applicable to dropbear
Typical dropbear configurations listen for all connections on port 22,
or filtered by the firewall to not listen on the WAN. If multiple
instances of dropbear are configured for different interfaces, changes
do not trigger a reload.
A good example is of openvpn - the interface is not up when dropbear
init is first called, so never listens for inbound connections from this
interface when it later comes up. This script detects if an interface
applicable to dropbear has come online and triggers a reload. If no
dropbear instance's interface is affected, no action is taken.